flux-resource: add -i, --include option to status and drain
Problem: There is no way to filter the output of the flux-resource(1)
status and drain commands to a set of included hosts or ranks.

Add a new option, -i, --include=TARGETS to these flux-resource commands
which will filter the list of presented results to a hostlist or idset
of ranks.

Fixes #5143