libfilmap: add library for common filemap routines
Problem: Code for handling extraction of RFC 37 File Archives is
duplicated between the flux-filemap(1) command and the shell stage-in
plugin.

Add a libfilemap convenience library which abstracts the filemap
extraction routines needed for both the flux-filemap and shell
stage-in use.

Support both a JSON array and JSON dictionary of file system objects,
with the path of each entry in a dictionary representation specified
by the key instead of 'path' in the fileref object