testsuite: fix potential flux-top race
Problem: The flux-top command can redraw output when a heartbeat
is received.  On a slow system, such as during a large parallel
run of the testsuite, its possible the heartbeat could arrive before
a test completes, resulting in unexpected output and a failing test.

Solution: In t2801-top-cmd.t, remove the heartbeat module so no heartbeats
can arrive to flux-top.

Fixes #5223