python: support add_file method for flux.job.Jobspec class
Problem: There is no way to add entries to the RFC 14 'files' attribute
in a Jobspec object.

Add a new 'add_file()' method for flux.job.Jobspec objects. The
method uses the Fileref class to add RFC 37 file objects to the
'files' dictionary in jobspec, and thus strings, bytes, and files
can be easily added to jobspec using this interface.