cli: add `--conf=CONF` option to flux-alloc,batch
Problem: There are no options to specify a config file for subinstances
created by `flux-batch(1)` and `flux-alloc(1)`, but it would be
extremely useful to be able to do so.

Add a `--conf=CONF` option to these commands. The option supports
several types of arguments, and may be specified multiple times to
update the built configuration:

 - If `CONF` is a multiline string, then parse `CONF` as JSON or
   TOML and update config.
 - If `CONF` contains an `=`, then parse `CONF` as config `KEY`=`VAL`
   and update config `KEY` with `VAL`, where `VAL` is parsed as JSON.
 - If `CONF` ends in `.json` or `.toml`, then treat `CONF` as a path to
   a config file to load as JSON or TOML.
 - Future extension: Otherwise, load a "named configuration" file from
   a local repository of config files.

The config is built iteratively using a new BatchConfig class in
cli/base.py