python: add --add-file option to cli submission commands
Problem: There is no easy way to add files to the Jobspec file archive
from the command line in the Flux submission commands.

Add a new `--add-file` option which allows a user to easily add a
file from the filesystem or data directly to a submitted jobspec. The
option takes the form `[NAME=]PATH` where the file will be encoded as
`NAME` if provided, otherwise the basename of `PATH` is used.

If PATH is a multiline string, then it is instead interpreted as the
file content to store, and is directly encoded as the data for file
NAME in Jobspec.