python: add flux.util.xdg_searchpath()
Problem: Creation of a search path based on the XDG standard may
be useful in multiple places, but the code to do this is currently
embedded in the UtilConfig class.

Move the construction of the XDG search path to its own function
for reusability.

Add an optional `subdir` parameter to the function so that callers
which need to build a set of paths in a subdir of the XDG search path
need not do this manually.