python: add filter method to SchedResourceList class
Problem: There is no way to filter the included ranks in a resource
set returned by the flux.resource.resource_list() method. This
would be useful in the implementation of a --include option for
flux-resource list.

Add a `filter` method to the SchedResourceList class so that the result
of resource_list() can be filtered by a include IDset or Hostlist.