cli: add --include option to `flux resource list` and `info`
Problem: An option to filter the output of `flux resource list`
would be useful, but nothing exists currently to do that.

Add an --include option to `flux resource list` that takes an IDset
of ranks or a hostlist to only include in the output.

Since `flux resource info` is a special case of `flux resource list`,
add the `-i, --include` option to that command as well.