python: do not call IDset(str(arg)) in ResourceSet class
Problem: A few places in the ResourceSet and Rlist classes convert
a ranks argument to a string before passing it to `IDset()`, but this
doesn't work for iterables like a set or list of integers. It is
also unnecessary since the IDset constructor already handles non-str,
iterable arguments.

Change `IDset(str(ranks))` to just `IDset(ranks)` where necessary.