completions: complete full subcommands
Problem: When the current argument matches a full subcommand name,
then the word is not completed with a tab. For example `flux top<TAB>`
will not add an extra space after `top`, even though this is the only
possible completion, even though `flux to<TAB>` would complete with
`flux top `. This is annoying.

The problem is that when a subcommand fully matches, the subcommand
helper function takes over, but since the subcommand "name" is the
current word, there are no possible completions.

Check if current matched any subcommands. If so, and if COMPREPLY is
empty, then complete the current subcommand.