python: add API to get job info from the KVS
Problem: There is not a convenient API to get job information
via the `job-info.lookup` RPC.

Solution: Add a new "kvslookup.py" module to the Python bindings.
It includes the "job_kvs_lookup()" function and the "JobKVSLookup"
class for users to retrieve data via the `job-info.lookup` RPC.
The module and interface is similar to the "list.py" module (
"get_job()" function and "JobList" class).

Fixes #5176