shell/pmi: make pmi1/2 aliases for simple
Problem: when a user specifies -o pmi=pmi2, the effect is
the same as -o pmi=off.

pmi methods that are not handled by any plugin are ignored.

Make pmi2 and pmi1 an alias for pmi2 so that the simple wire
protocol is offered in both cases.  The flux libpmi2.so actually
uses the simple wire protocol.  Slurm's libpmi2.so _could_ use it
but due to a bug fails version negotiation.

So maybe this is better than ignoring it.

Fixes #5226.