flux-resource: abstract fetch of resource list
Problem: The fetch of reosurce list and config may be a common
operation in the flux-resource command, but is currently embedded in
the list command function list_handler().

Add a `get_resource_list()` function that combines the common
operations of fetching the scheduler resource list and instance config,
processing the --states and --include options.