flux-resource: add `R` subcommand
Problem: There is no way for a user to easily get the R being used
for job allocations in an instance. The resource.R KVS key is only
available to the instance owner, and doesn't represent the scheduler
view of resources.

Add a `flux resource R` command that fetches the scheduler's operating
resource set and emits an R on stdout. Similar options to `flux
resource list` are supported so that a user can get an R in any of
the supported states "up", "down", "allocated", "free", "all".