shell: use f58 jobid encoding for shell service
Problem: The shell service name uses the decimal encoding of the
jobid as the suffix, but using the f58 encoding would make it more
clear that the service corresponds directly to a job.

Encode the jobid in the shell service name as f58.

Since the job-exec module builds the shell service topic string by
hand, update it as well.