liboptparse: add optparse_get_size() accessor
Problem: It would be convenient to easily support "size" option
arguments, e.g. 1M 250k, .5G,  with liboptparse, but there is no
custom accessor for arguments that supports this.

Add optparse_get_size() which supports parsing a string argument
which specifies a file or other size in bytes, with optional suffix
'b', 'k', 'M', or 'G'.