flux-top: fix formatting with ASCII jobids
Problem: When jobids are being encoded with ascii 'f', jobs in
the flux-top(1) joblist pane are not aligned with the header row.
This occurs because mvwprintw(3) doesn't seem to handle multibyte
UTF-8 characters properly. (This appears to be why the JOBID column
is formatted with %12s while the jobs are formatted with %13.13s
currently).

Detect whether ascii 'f' is being used in jobids by testing the result
of strlen(3) for jobid 0 encoded as f58. If the result is 2 then
ascii jobids are being used, so use 13 for the JOBID column width,
otherwise the multibyte 'ƒ' is most likely being used so use 12 for
the JOBID width.