python: Add host_ranks method for ResourceSet objects
Problem: It is tricky to map a set of hostnames to their broker ranks
given a ResourceSet object. Though all the information is present,
you first have to find the index of the desired hosts in the resource
set nodelist, then use those indices to find the corresponding ranks
in the ranks IDset.

Add a `host_ranks` method to ResourceSet to make this operation simple
for users. By default a FileNotFound exception is raised if any hosts
are not found in the ResourceSet, but an ignore_nomatch option is
provided to just return the ranks of matching hosts.