testsuite: test ResourceSet.host_ranks() method
Problem: There are no unit tests of the ResourceSet.host_ranks()
method in the testsuite. Additionally, the ResourceSet tests do not
test an Rv1 with non-zero origin set of ranks.

Add a secondary Rv1 test object to t/python/t0022-resource-set.py
with a set of nonzero origin ranks for testing, and add tests for
the host_ranks() method.