python: fix filter() method in SchedResourceList class
Problem: The filter() method in the SchedResourceList class does not
properly derive the set of ranks given a set of hosts. This is because
the implementation assumes the index into the nodelist corresponds
directly to broker ranks, but these indices need to then be used to
find the broker ranks in the `ranks` attribute.

Now that the ResourceSet class exports a host_ranks method, use
that instead to fix the incorrect derivation of broker ranks and fix
the bug.

Fixes #5266