shell: add pty 'capture' member to shell.init context
Problem: If a user launches an interactive instance (i.e. flux alloc)
with `-o pty.capture`, then data from the pty is duplicated into
the output eventlog for later use. However, there is no way for an
end user (e.g. flux job attach) to know definitively if that data
is duplicated, because there is no way to easily discover (besides
fetching the original jobspec) if capture mode was enabled.

If capture mode is enabled with an interactive pty, add `"capture":1`
to the shell.init event. This will give flux-job the hint it needs to
determine that data in the output eventlog for rank 0 is duplicated
from the pty output.