flux-job: handle log and output messages with pty attach
Problem: When a job uses an interactive pty, such as with
flux-alloc(1), the job's output eventlog is ignored because it is
presumed all output is being handled through the pty. However,
this means that all log messages, including error messages, are
suppressed. This can result in confusing behavior. For instance,
if the shell cannot change its working directory, it logs an error
message. If the user does not see this message they may think the
software is broken for other reasons.

Process the output eventlog with `flux job attach` even when
interactively attaching to a pty. Task output and log messages are
sent the users original terminal. Since the terminal is assumed to be
in raw mode, add the missing carriage return ('\r') to these messages
to avoid leaving the cursor in the wrong column. If interactive pty
output is also being captured via the pty.capture option, then skip
output for rank 0 since it will likely be duplicated.

Fixes #5272