flux-filemap: use optparse_get_size_int() for size options
Problem: The flux-filemap(1) --chunksize and --small-file-threshold
options take arguments in bytes, but do not support common suffixes
like k, K, M, or G. This makes the options more cumbersome to use
than necessary.

Use optparse_get_size_int() to fetch the value of these options,
which allows the argument to be specified as a floating point number
with optional suffix in the set [kKMG] (up to but not including 2G).