libsubprocess: add flux_subprocess_channel_incref/decref
Problem: It would be useful to take a reference on an output stream
of a subprocess so that stream doesn't generate EOF until all references
have been dropped. This can be used when multiple sources are being
spliced onto a subprocess output to ensure that the subprocess doesn't
register completion until the other outputs have also been fully read.

Add flux_subprocess_channel_incref/decref which wrap the underlying
flux_buffer_read_watcher_incref/decref for the channel in question.