pty: send EOF to monitor on EIO
Problem: The pty monitor callback is not called when the pty "follower"
has closed, i.e. all open handles using the pty have closed. However,
this would be extremely useful information to share with the monitor.

When an EIO error is receieved on the pty, send an EOF message to the
monitor (len == 0) to indicate there are no more pty users.