shell: pty: prevent eof before all pty data is captured
Problem: When the pty.capture shell option is used (either implicitly
or explicitly), an eof for stdout may be written to the output eventlog
before all pty data is read. This is particularly true when used with
pty.interactive, since in this mode the normal stdout file descriptor
for the task is closed immediately when the task is attached to
the pty.

When capturing pty output for the output eventlog, take a reference
on the task's stdout channel so that it doesn't close until the pty
has completed reading its data.