testsuite: ensure eof is the last output event with pty.capture
Problem: The shell pty tests do not test that eof is the last event
in the output eventlog when using pty.capture.

Enhance the test in t2612-job-shell-pty.t that runs with
pty.interactive and pty.catpure to run multiple times and ensure that
the eof is the final output eventlog event in all cases.