flux-job: attach: avoid watching output eventlog twice
Problem: There is a potential for flux-job to call the function
attach_output_start() twice when pty.capture and pty.interactive are
enabled for the job. This causes the output eventlog to be watched
twice and duplicates lines of ouput.

Make attach_output_start() idempotent, i.e. don't do anything if the
output eventlog is already being watched.