testsuite: fix occasional broker kill error
Problem: flux-start will kill all brokers after the first broker
exits.  By default the timeout to kill all the other brokers
is 20s.  A broker in a test in t3203-instance-recovery.t would
sometimes be killed due to a slow running rc3 cleanup.  This would
lead to a invalid test error.

Solution: Increase the kill timeout to 300s with the--test-exit-timeout
option.

Fixes #5290