libutil/fluid: add FLUID_STRING_F58_PLAIN
Problem: in non-UX sitations it may be safer or required
to encode FLUIDs using a prefix of f not ƒ, but there
is no convenient way to do that.

Add FLUID_STRING_F58_PLAIN encoding, which is the same as
FLUID_STRING_F58, except that the prefix is forced to f.

Update unit test.