flux-shell: use f58plain in shell service name
Problem: the job-exec module reconstructs the shell's service
name to send the rank 0 shell an exception message when one of
the other shells goes offline, but when f58 encoding is used,
the reconstruction could be different when the shell node and
the rank 0 node have different locale configurations.

Use the f58plain jobid encoding to avoid this potential mismatch.

Fixes #5257