libjob: export flux_unwrap_string(3) function
Problem: The internal libjob function unwrap_string() allows
users to unwrap signed data regardless if flux-core was built
with flux-security or not.  It could be useful for libraries/tools
that are not a part of flux-core.

Solution: Rename unwrap_string() to flux_unwrap_string() and make it
a public function.  Update unit tests and callers accordingly.  Remove
include of unwrap.h where it is no longer needed.