python: support "original" in job kvslookup module
Problem: The "flux job info" command has a "--original" option to
get the "original" jobspec for a job.  It would be nice if there
was a similar option in the job kvslookup python module.

Solution: Add a "original" option to job_kvs_lookup() and JobKVSLookup.
Internally it will lookup "J" and decode the original jobspec from it.

Fixes #5286