libflux: do not force next future to be fulfilled every time
Problem: The implementation of "chained" futures, i.e. those created
with flux_future_and_then(3) or flux_future_or_then(3), currently
assumes that the 'next' future in the chain will either be fulfilled
explicitly from within the callback, or implicitly by forcing an
unfulfilled future to be fulfilled by the current future in the
chained_continuation() callback. This prevents one use case for an
and_then() future though: fulfilling the next future with only a subset
of values from another multiple fulfillment future. For example,
one such case is watching an output eventlog and only processing
'data' events.

Check if the previous future has been reset in the chained future
continuation and immediately return if so. This most likely indicates
that the next future was purposefully unfulfilled, and will be
fulfilled instead by a subsequent callback.