job-info: ignore duplicate lookup keys
Problem: When doing a job lookup, if a lookup requests the same key
twice, job-info returns a not very obvious what is wrong EEXIST
error to the user.

Solution: Ignore duplicate keys.  Since the response to a lookup
request is a json object, duplicate keys would not lead to a different
result (i.e. you can't add the same key twice to a json object).

Fixes #5303