flux-job: fix invalid --original info output
Problem: In flux job info and the --original option, it is assumed
that every response of "J" is meant to be "jobspec".  However, that
is not a safe assumption, as the user may have input both "J" and
"jobspec".

Solution: When outputting the results from a flux job info lookup,
output results based on the keys input by the user, not by the keys
used during lookup.

Fixes #5305