job-manager: do not terminate epilog on job exception
Problem: The job-manager 'perilog' plugin terminates any running prolog
if the job receives an exception. This works as expected, but the code
doesn't check whether a running process is the prolog or epilog, so
job epilogs are killed on job exception as well. This isn't desired
behavior since this doesn't allow the epilog to properly complete,
and also causes the nodes on which the epilog is running to be drained.

Ensure that the running process is a prolog before killing it.

Fixes #5319