python: fix writes in KVSDir with initial paths
Problem: KVSDir can be passed an initial path when initialized.  This
path gives context to all reads.  For example, if initialized with the
path "foo", a later read of the key "bar" will actually read "foo.bar".
While this works for reads, this does not work for writes.  For example,
if initialized with the path "foo", a later write of the key "bar" will
write to the path "bar", not "foo.bar".

Solution: When doing writes, ensure writes are written relative to the
initial path.

Fixes #5308