python: raise handle exception after blocking Future calls
Problem: When a Python callback is called as part of a blocking Future
get() or wait_for() call and that callback raises any exception, this
could result in the underlying `flux_future_t` being fulfilled with
an error.  Upon returning to Python, the original exception that caused
the error is masked because Future.get()/wait_for() and derivatives
will raise the OSError instead of the exception stored in the handle.

A specific case where this occurs is when the Future in question
is supposed to be fulfilled as a result of the Python callback,
but the exception occurs before that can happen. In this case,
flux_future_wait_for(3) returns with EDEADLOCK since this function
should block until the future is fulfilled. Raising OSError with
errno=EDEADLOCK is not helpful, though. Instead the user should see
the exception which is stored in the handle.

Add a new Future method `raise_if_handle_exception()`, which checks for
any handle exception and raises it. Call this function on return from
any low-level, blocking Future calls, including Future.wait_for(),
Future.get(), JobEventWatchFuture.get_event(), RPC.get_str(), and
JobResultFuture.get_dict().