python: add FutureExt class
Problem: There is no way to construct Flux Future objects that can
be initialized and fulfilled from Python.

Add the FutureExt class which allows empty futures with an initializer
callback to be constructed from Python. This class supports a fulfill()
method which allows the Future to be fulfilled with serializable
Python objects.