python: make EventLogEvent a subclass of dict
Problem: There is no way to retrieve the original dict used to create
an instance of the EventLogEvent class, but this will be useful in many
situations. Also, it will be very useful to have the EventLogEvent objects
be JSON serializable.

Make the EventLogEvent class a subclass of dict. This not only allows the
original entry dict available as dict(event), but also makes objects of
this class JSON serializable. Clean up object initialization by just keeping
the attributes in the event entry dict.