python: add error string when re-raising get_event() exceptions
Problem: When calling the ffi RAW.get_event() wrapper in the
JobEventFuture class, errors set on the underlying flux_future_t object
are translated to an OSError using the set errno. When this occurs,
any error string associated with the future is lost.

When handling an error from JobEventFuture.get_event() copy the future
error string over exception strerror so this extra detail is available
to the exception handler.