python: always raise pending handle exception
Problem: When a nested invocation of the reactor on a handle is stopped
with an error, sometimes the original reactor does not exit with rc < 0,
so the exception is lost.

Since there should not be an exception registered with the flux.Flux
class unless an unhandled exception is generated in a callback, drop
the check for `rc < 0` and always call `Flux.raise_if_exception()`
after leaving the reactor. This should result in all exceptions that
occur from within Python under flux_reactor_run(3) to be raised instead
of silently dropping some exceptions (making noisy bugs instead of
silent ones).