testsuite: order python kvs tests
Problem: The tests in python/t0005-kvs.py were not numbered, so
when running only the kvs tests the tests came out ordered
lexicographically, which had the appearance of random output order.

Add numbers to tests so they come out in expected order given their
placement in the test file.