python: clear KVS txns on commit error
Problem: When a commit error happens on a KVS transaction, an error
is raised.  This means that the internal KVS transaction (created
via calls to put(), put_mkdir(), etc.) does not get cleared.  Therefore
a subsequent call the commit() would redo the the transaction.  If
the transaction is the cause of the commit failure, no subsequent
commits can ever work.

Solution: If an exeption is raised during the commit, clear the internal
transactions before returning the exception to the user.

Fixes #5333