resource: handle exclusion with dynamic R
Problem: when R is acquired dynamically, the resource module refuses
to process exclusions.

inventory_targets_to_ranks() fails if R is not set, but _R_ is only
used if the input fails to parse as an idset.

Move the check for NULL R into the hostnames to ranks mapping block
where it is used.  This allows an exclusion set specified as an idset
to be decoded when R is dynamically discovered.

Also ensure that the flux_error_t parameter is filled on all error paths.

Fixes #5337