job-manager: always output an exception note
Problem: Currently when an exception is raised, the job-manager
will output an empty string if an exception note is not specified.
Code indicates that this is a workaround due to the fact an exception
note is optional.

An update to RFC21 makes this exception note a requirement.  Therefore
the workaround is no longer needed.

Solution: Always output an exception note.  If one is not provided by the
user, output a note of empty string.

Fixes #5134