python: exception events require notes
Problem: In the JobException class, events are assumed to optionally
have notes.  An RFC21 update makes notes a requirement.

Solution: Do not assume a note is optional, assume one always exists.

Note that a bug / feature in job-manager previously set an exception note
to empty string even if there was not one.  So this should not break on
eventlogs in older versions of flux.