flux-job: parse required exception note
Problem: An exception is parsed in flux-job attach as though the exception
note is optional.  A recent RFC21 update makes the note a requirement.

Solution: Do not parse the exception note optionally, always parse it.

Note that a bug / feature in job-manager previously set an exception note
to empty string even if there was not one.  So this should not break on
eventlogs in older versions of flux.