job manager: parse exception note as requirement
Problem: In the job-manager wait code an exception note is parsed
optionally.  An RFC21 update makes an exception note a requirement.

Solution: Do not parse the exception note optionally, parse it as a
requirement of the exception.

Note that a bug / feature in job-manager previously set an exception note
to empty string even if there was not one.  So this should not break on
eventlogs in older versions of flux.