job-list: exception note is required
Problem: job-list parses an exception note as an optional field.  Per
an RFC21 update, it is no longer optional.

Solution: If an exception is parsed from the eventlog, always parse the exception
note.

Note that a bug / feature in job-manager previously set an exception note
to empty string even if there was not one.  So this should not break on
eventlogs in older versions of flux.