libjob: do not parse note optionally
Problem: The note in an exception event is no longer optional
per an RFC21 update.

Solution: Do not parse the exception note optionally.

Note that a bug / feature in job-manager previously set an exception note
to empty string even if there was not one.  So this should not break on
eventlogs in older versions of flux.