job-manager: require start response note
Problem: The <exec-service>.start response can contain an exception,
with an optional note.  RFC21 was recently updated that eventlog
exception notes are required.  While this does not affect the start
response, it would be nice if things were consistent.

Solution: For consistency, make the exception note in a start
response a requirement.  Parse it as a requirement in the job manager.