python: fix potential abort with multiply fulfilled futures
Problem: If a future does not call future.reset() before exiting its
then_cb, then it is not only deleted from the global _THEN_HANDLES
dict, but also future.cb_handle is set to None. This allows potential
garbage collection of the ffi handle, which will cause a Python
abort if the future is still fulfilled or fulfilled again and enters
continuation_callback() after the ffi handle is GC'd.

To avoid this problem, do not set future.cb_handle = None when the
counter in _THEN_HANDLES[future] drops to zero. Instead, let the ffi
handle lifetime be the same as the Future itself. Once the Future is
destroyed, the underlying `flux_future_t` will be too, so there is
no danger of the continuation_callback() being run after this occurs.

Since a Future may now exist without being in the _THEN_HANDLES
dict, check that the future is in that dict before operating on it
in continuation_callback() and future.reset().