python: stop futures if callback raises exception
Problem: An exception raised in a future's continuation callback may
leave the future fulfilled and active. If the reactor is restarted,
then the callback could be called again, and the same exception raised.

Reset and stop any future that raises an exception in their callback.
This should avoid the problematic callback from being re-invoked.