python: add JobWatcher class
Problem: There is no convenient way to watch or monitor multiple jobs
simultaneously in the Python API.

Add a new module `flux.job.watcher` with a `JobWatcher` class that
can be used to watch and monitor the progress of one or many jobs.
This class is meant to replace the functionality currently embedded
in the `flux submit --watch` and `flux bulksubmit --watch` commands,
and thus supports a similar set of options. However, this class could
be used to "watch" jobs that are already running and/or complete.