testsuite: ensure flux-submit --dry-run works without Flux
Problem: No tests in the testsuite ensure that flux-submit with
the --dry-run option works without a running Flux instance, but this
case is expected to work by many external test suites.

Add a check to t2710-python-cli-submit.t that ensures

 flux submit -n1 --dry-run hostname

works when FLUX_URI is set to a non-existent path.