job-info: support WAITCREATE on eventlog watch
Problem: Sometimes a path a user wants to watch via
the job-info.eventlog-watch service may not yet be created
when the watch begins.  This leads to a ENOENT error.

This is most notable in guest namespaces, where the shell
is responsible for creating the path, and there can be
raciness between the start of the job (i.e. creation of the
guest namespace) and when the path is created.

Solution: Support a flag FLUX_JOB_EVENT_WATCH_WAITCREATE
to wait for the path to be created.  The flag will set
the FLUX_KVS_WAITCREATE underneath the covers as needed to
wait for the path to be created.

Fixes #5301