job-manager: epilog-start should prevent clean event
Problem: An epilog-start event emitted for a job that never had an
alloc event will proceed immediately to inactive because there is no
pending "free" event to hold the job until the matching epilog-finish
is posted.

A pending epilog-start event should prevent the "clean" event for
a job as well as the "free" event. This means a job cannot become
inactive until all epilogs are complete.

Fixes #5342