flux-job: suppress attach status line in more situations
Problem: The `flux job attach` "statusline" is really only useful when
the command is definitely being run interactively. In other situations,
the output could be quite annoying since the status line is updated
once per second. This could lead to a lot of output for a `flux run`
that is redirected to an output file, for example.

Suppress the statusline unless all of stdout, stderr, and stdin
are connected to a tty. Also, add a new FLUX_ATTACH_NONINTERACTIVE
environment variable that can be used to hint that `flux job attach`
is not running interactively, even if under a pty.

Since the complexity of the statusline initialization is now increased,
move the code to a new function.

Fixes #5343