flux-job: attach: always exit nonzero on fatal job exception
Problem: If a job finishes with 0 exit status, `flux job attach`
will also exit with a 0 exit status even if the job received a fatal
exception before the finish event. The fatal job exception should
result in a nonzero exit status even if all job tasks completed
successfully.

Track if a job got a fatal job exception and ensure that the
cmd_attach() returns nonzero if so.