job-list: apply FAILED result to jobs with any fatal exception
Problem: The job-list module marks jobs that finish with a status of 0
as successful, even if a fatal exception was posted to the job. This
is not consistent with other reporters of job status, e.g. `flux job
wait`, `flux job result` and `flux job attach`.

Only mark a job as "successful" if it finished with 0 wait status
and did not have a severity=0 job exception.