testsuite: ensure commands agree on status of job with exception
Problem: There is no testing that all utilities that report a job
status or result agree that any job with a fatal job exception has
failed, even if the job finished with a 0 wait status.

Since it is convenient that t2304-sched-simple-alloc-check.t reproduces
the conditions for this test (fatal job exception with zero exit
stauts), test that the job with the alloc-check exception is marked
as failed by a few other tools besides `flux job wait`.