flux-submit: exit with nonzero exit status for fatal exception
Problem: flux-submit and bulksubmit may exit with status 0 when
the job tasks finish with status 0, even if there is was fatal job
exception. This is inconsistent with the behavior of other utiliites
like `flux job wait`, `flux job status`, and `flux job attach`.

Always exit with exit code 1 if a fatal job exception is received by
a job, even if the job tasks finish successfully.

Update one test that assumes `flux submit` will exit with success in
this case.