shell: gpubind: use cpuset array for per-task support
Problem: The gpubind plugin "per-task" works by allocating gpu ids
from a shared gpus idset within the task.init callback, but this is
different from how the cpu affinity plugins and thus code can't be
easily shared.

Rewrite the gpubind plugin to use an array of hwloc_cpuset_t objects
to maintain which gpus should be assinged to each task. To ease memory
management, add a context object that contains relevant data for the
plugin, and create this object at initialization.

The gpubind plugin now assigns GPUs to tasks during initialization,
as does the affinity plugin. This allows the cpuset array code to be
reused, and for the gpubind plugin to support a "map:LIST" option in
the future.