shell: affinity: do not error if mapped cpuset outside of job
Problem: The usefulness of the `-o cpu-affinity=map:LIST` option is
reduced because the assigned cores must be contained within the cpu
set assigned to the job. However, if a user is using the `map:LIST`
option then they are presumably agreeing to take complete control of
core assignments, and the option should not limit them to only the
assigned cores.

Drop the check in the affinity plugin that cores are contained within
the job cpuset. Don't even check if the cores are valid (this would
presumably fail later when the affinity is actually applied.)

Fixes #5352