shell: gpubind: don't exit early from plugin if no gpus assigned
Problem: The gpubind plugin returns early if no GPUs were allocated
to the job, but some future gpu-affinity options may want to override
the default binding in this case.

Let the plugin fall through to the if/else block even if no GPUs are
assigned to the job. Change the final `else` to `else if (ngpus >
0)` though, so that the plugin still does nothing (besides setting
CUDA_VISIBLE_DEVICES=-1) if ngpus == 0.