shell: gpubind: support gpu-affinity=map:LIST
Problem: The gpubind shell plugin doesn't support explicit mapping
of GPUs to tasks.

Support a `-o gpu-affinity=map:LIST` which works similar to
the cpu-affinity of the same name. This option allows explicity
specification of GPUs to tasks without regard for the actual GPU ids
assigned to a job. It is mainly meant for testing, benchmarks, or when
default GPU assignment is not working for a particular situation.

Fixes #5350