python: fix job returncode for failed but not canceled jobs
Problem: A job that receives a fatal exception before starting could
end up with a returncode of an empty string because only canceled
jobs are handled before the RUN state.

Check for a FAILED job with no waitstatus when computing the JobInfo
returncode, and set the returncode to 1 in this case so an empty
returncode is not returned.