shell/oom: plugin to report oom kills
Problem: when cgroup memory limits are set and a job task
is killed by the kernel oom killer, the user is not told that
they exceeded a limit.

Add a simple shell plugin that watches the shell's cgroup memory.events
file with inotify(7) and reports on the job's stderr when an oom kill
occurs.

Fixes #5361