job-list: support legacy RPC protocol
Problem: It would not be unreasonable for a user to launch a flux
version that is different than the parent version of flux (especially
when then parent version is the system instance).  With the recent
update of the job list module to use "constraints", the older RPC
format would not allow a user to (for example) run:

flux --parent jobs

If the parent version is a newer version of flux.  Or run:

flux jobs -R

if the subinstance is a newer version of flux.

Solution: Support the older RPC protocol in job-list.  Internally,
simply create a constraint object given the old protocol's inputs.

Fixes #5363