testsuite: fix shell setopt test
Problem: the shell setopt test is sensitive to the order that builtin
plugins are initialized.

Specifically, the "flux-shell: plugins can use setopt with empty options"
test in t2603-job-shell-initrc.t fails when the oom plugin is initialized
after the affinity plugin.

The setopt.c test plugin expects that flux_shell_info_unpack() sets
errno = ENOENT on failure, but this is not the case and the test is
likely relying on a residual errno value from one of the builtin plugin
init callbacks.

Fix the test to just check for failure and not expect errno to be set.